% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{calibration_stats}
\alias{calibration_stats}
\title{Computes calibration alpha and slope from Cox regression}
\usage{
calibration_stats(cox_model, testdata)
}
\arguments{
\item{cox_model}{fitted cox model}

\item{testdata}{test data}
}
\value{
c("calib_alpha"=calib_alpha, "calib_slope"=calib_slope)
}
\description{
Details: function computes for a given coxph() model and test
data calibration alpha and slope, as per
}
\section{https://journals.sagepub.com/doi/pdf/10.1177/0962280213497434}{
Crowson, C. S., Atkinson, E. J., & Therneau, T. M. (2016).
Assessing calibration of prognostic risk scores.
Statistical methods in medical research, 25(4), 1692-1706.
}

