% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cox_calibration_stats}
\alias{cox_calibration_stats}
\title{Calibration stats of a fitted Cox PH model}
\usage{
cox_calibration_stats(cox_model, test_data)
}
\arguments{
\item{cox_model}{fitted cox model, namely, coxph() object}

\item{test_data}{test data, should be a data frame with "time" and "event" columns for survival outcome}
}
\value{
c(calibration alpha, calibration slope)
}
\description{
Computes calibration alpha and slope for a fitted coxph model
in the data.

Crowson, C. S., Atkinson, E. J., & Therneau, T. M. (2016).
Assessing calibration of prognostic risk scores.
Statistical methods in medical research, 25(4), 1692-1706.

https://journals.sagepub.com/doi/pdf/10.1177/0962280213497434
}
