% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survival.R
\name{df_event_times}
\alias{df_event_times}
\title{Auxiliary function for simsurve_linear, non_linear, crossterms}
\usage{
df_event_times(
  exp_beta,
  df,
  N,
  observe_time,
  percentcensored,
  randomseed,
  lambda,
  distr,
  rho_w,
  drop_out
)
}
\arguments{
\item{exp_beta}{exponential distribution beta param}

\item{df}{data}

\item{N}{sample size}

\item{observe_time}{observation time}

\item{percentcensored}{percent of censored observations (with no event) by observe_time}

\item{randomseed}{random seed}

\item{lambda}{exponential/Weibull distribution lambda param}

\item{distr}{distribution for times, "Exp" or "W" for exponential or Weibull}

\item{rho_w}{Rho parameter for Weibull distribution}

\item{drop_out}{drop out rate of those censored before observe_time}
}
\value{
data frame with added columns "time", "event", "event_time"
}
\description{
Auxiliary function for simsurve_linear, non_linear, crossterms
}
