% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcompare.R
\name{difftest}
\alias{difftest}
\title{Testing statistical significance of the ensembled model outperformance
over the baseline model (Model 1 vs Model 0)}
\usage{
difftest(res1, res0, sample_n, param_n)
}
\arguments{
\item{res1}{data frame}

\item{res0}{list of predictors to use}

\item{sample_n}{prediction time of interest. If NULL, 90\% quantile of event times is used}

\item{param_n}{random seed for replication}
}
\value{
matrix: mean, std and p-values for AUC, BS, BS_scaled, C_score, Calib_slope, Calib_alpha
}
\description{
Testing statistical significance of the outperformance of
Model 1 over Model 0
T-test is used for AUC and C-score, and Fisher test for Brier Scores
Models are given with the results of the survval function applied
to several cross-validation or bootstrap iterations, one per row
It is assumed that the same training data is used for Model 1 and Model 2
in each row
}
