% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_cox_coxlasso.R
\name{eligible_params}
\alias{eligible_params}
\title{Internal function, checks eligible parameters
to avoid tree-building with non-existing or constant predictors}
\usage{
eligible_params(params, df)
}
\arguments{
\item{params}{list of predictor names}

\item{df}{data frame}
}
\value{
params_eligible - list of params which are in the df and have more than 1 value
}
\description{
Internal function, checks eligible parameters
to avoid tree-building with non-existing or constant predictors
}
