% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{linear_beta}
\alias{linear_beta}
\title{Auxiliary function for simulatedata functions}
\usage{
linear_beta(df)
}
\arguments{
\item{df}{data}
}
\description{
Auxiliary function for simulatedata functions
}
