% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDeepHit.R
\name{ml_hyperparams_deephit}
\alias{ml_hyperparams_deephit}
\title{Internal function for getting grid of hyperparameters
for random or grid search of size = max_grid_size}
\usage{
ml_hyperparams_deephit(
  mlparams = list(),
  max_grid_size = 10,
  dftune_size = 1000,
  randomseed = NaN
)
}
\arguments{
\item{mlparams}{list (mod_alpha,dropout, learning_rate, num_nodes, epochs, sigma, cuts,batch_size, early_stopping, weight_decay)}

\item{max_grid_size}{number of random grid searches for model tuning}

\item{dftune_size}{size of the tuning data to define batch size options}

\item{randomseed}{random seed to control which rows from the grid are selected}
}
\description{
Internal function for getting grid of hyperparameters
for random or grid search of size = max_grid_size
}
