% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survival.R
\name{nonlinear_beta}
\alias{nonlinear_beta}
\title{Auxiliary function for simulatedata functions}
\usage{
nonlinear_beta(df)
}
\arguments{
\item{df}{data}
}
\description{
Auxiliary function for simulatedata functions
}
