% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_1.R
\name{predict.survensemble1}
\alias{predict.survensemble1}
\title{Predicts event probability for a fitted Ensemble 1 method}
\usage{
\method{predict}{survensemble1}(survens1, df_test, fixed_time, oob = FALSE, useCoxLasso = FALSE)
}
\arguments{
\item{survens1}{trained survensemble1 model}

\item{df_test}{test data}

\item{fixed_time}{time for which probabilities are computed}

\item{oob}{TRUE/FALSE , default is FALSE}

\item{useCoxLasso}{TRUE/FALSE , default is FALSE}
}
\value{
returns matrix(nrow = length(newdata), ncol = length(times))
}
\description{
\link[survcompare:predict.survensemble1]{predict.survensemble1}
}
\examples{
df <- simsurv_nonlinear()
}
