% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survcompare.R
\name{print.survcompare}
\alias{print.survcompare}
\title{Print survcompare object}
\usage{
\method{print}{survcompare}(x, ...)
}
\arguments{
\item{x}{output object of the survcompare function}

\item{...}{additional arguments to be passed}
}
\value{
x
}
\description{
Print survcompare object
}
