% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ensemble.R
\name{print.survensemble}
\alias{print.survensemble}
\title{Prints trained survensemble object}
\usage{
\method{print}{survensemble}(x, ...)
}
\arguments{
\item{x}{survensemble object}

\item{...}{additional arguments to be passed}
}
\value{
x
}
\description{
Prints trained survensemble object
}
