% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_1.R
\name{print.survensemble1}
\alias{print.survensemble1}
\title{prints survensemble1 object}
\usage{
\method{print}{survensemble1}(obj, ...)
}
\arguments{
\item{obj1}{survensemble1 object}
}
\description{
prints survensemble1 object
}
