% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_print.R
\name{print.survensemble_cv}
\alias{print.survensemble_cv}
\title{Prints trained survensemble object}
\usage{
\method{print}{survensemble_cv}(x, ...)

\method{print}{survensemble_cv}(x, ...)
}
\arguments{
\item{x}{survensemble_cv object}

\item{...}{additional arguments to be passed}
}
\value{
x

x
}
\description{
Prints trained survensemble object

Prints survensemble_cv object
}
