% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survensemble.R
\name{print.survensembleCV}
\alias{print.survensembleCV}
\title{Prints survensembleCV object}
\usage{
\method{print}{survensembleCV}(x, ...)
}
\arguments{
\item{x}{survensembleCV object}

\item{...}{additional arguments to be passed}
}
\value{
x
}
\description{
Prints survensembleCV object
}
