% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survival.R
\name{simsurv_lin_nonPH}
\alias{simsurv_lin_nonPH}
\title{Simulated sample with survival outcomes with non-linear and cross-term dependencies and non-PH survival times.}
\usage{
simsurv_lin_nonPH(
  N = 1000,
  observe_time = 10,
  percentcensored = 0.75,
  randomseed = NULL,
  lambda = 0.1,
  drop_out = 0.3
)
}
\arguments{
\item{N}{sample size, 1000 by default}

\item{observe_time}{study's observation time, 10 by default}

\item{percentcensored}{expected number of non-events by observe_time, 0.75 by default (i.e. event rate is 0.25)}

\item{randomseed}{random seed for replication}

\item{lambda}{baseline hazard rate, 0.1 by default}

\item{drop_out}{expected rate of drop out before observe_time, 0.3 by default}
}
\value{
data frame; "time" and "event" columns describe survival outcome; predictors are "age", "sex", "hyp", "bmi"
}
\description{
Simulated sample with Weibull distributed time-to-event;
log-hazard are non-linear and with cross-terms. Survival curves for different sexes intersect (non-PH).
}
\examples{
mydata <- simsurv_lin_nonPH()
head(mydata)
}
