% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survival.R
\name{simsurv_linear}
\alias{simsurv_linear}
\title{Simulates population with linear dependencies}
\usage{
simsurv_linear(
  N = 1000,
  observe_time = 10,
  percentcensored = 0.75,
  randomseed = NULL,
  lambda = 0.1,
  distr = "Exp",
  rho_w = 1,
  drop_out = 0.3
)
}
\arguments{
\item{N}{sample size}

\item{observe_time}{study's observation time, default is 10}

\item{percentcensored}{expected number of non-events by observe_time, 0.75 by default (i.e. event rate is 0.25)}

\item{randomseed}{random seed for replication}

\item{lambda}{baseline hazard rate, 0.1 by default}

\item{distr}{time-to-event distribution, "Exp" for exponential (default), "W" for Weibull}

\item{rho_w}{shape parameter for Weibull distribution, 0.3 by default}

\item{drop_out}{expected rate of drop out before observe_time, 0.3 by default}
}
\value{
data frame with columns "age", "sex", "hyp", "bmi", "time", "event", "event_time"
}
\description{
Simulates population with linear dependencies
}
\examples{
mydata <- simsurv_linear(1000)
head(mydata)
}
