% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survival.R
\name{simulate_population}
\alias{simulate_population}
\title{Auxiliary function for simulatedata}
\usage{
simulate_population(N = 1000, randomseed = 42)
}
\arguments{
\item{N}{sample size}

\item{randomseed}{random seed, 42 by default}
}
\value{
data frame
}
\description{
Simulates age, bmi, hyp and sex for a generated population
}
