% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_survivalrandomforest.R
\name{srf_survival_prob_for_time}
\alias{srf_survival_prob_for_time}
\title{Internal function to compute survival probability
by time from a fitted survival random forest}
\usage{
srf_survival_prob_for_time(rfmodel, df_to_predict, fixed_time, oob = FALSE)
}
\arguments{
\item{rfmodel}{pre-trained survrf_train model}

\item{df_to_predict}{test data}

\item{fixed_time}{at which event probabilities are computed}

\item{oob}{TRUE/FALSE use out-of-bag prediction}
}
\value{
output list: output$train, test, testaverage, traintaverage, time
}
\description{
Internal function to compute survival probability
by time from a fitted survival random forest
}
\examples{
df <- simsurv_nonlinear()
#params<- c("age", "hyp", "bmi")
#s <- survrf_train(df, params)
#p <- survrf_predict(s, df, 5)
}
