% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survcompare.R
\name{summary.survcompare}
\alias{summary.survcompare}
\title{Summary of survcompare results}
\usage{
\method{summary}{survcompare}(object, ...)
}
\arguments{
\item{object}{output object of the survcompare function}

\item{...}{additional arguments to be passed}
}
\value{
object
}
\description{
Summary of survcompare results
}
