% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ensemble.R
\name{summary.survensemble}
\alias{summary.survensemble}
\title{Prints summary of a trained survensemble object}
\usage{
\method{summary}{survensemble}(object, ...)
}
\arguments{
\item{object}{survensemble object}

\item{...}{additional arguments to be passed}
}
\value{
object
}
\description{
Prints summary of a trained survensemble object
}
