% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_1.R
\name{summary.survensemble1}
\alias{summary.survensemble1}
\title{summary for a trained Cox-SRF ensemble, should be a survensemble1 object}
\usage{
\method{summary}{survensemble1}(obj1)
}
\arguments{
\item{obj1}{survensemble1 object}
}
\description{
summary for a trained Cox-SRF ensemble, should be a survensemble1 object
}
