% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ensemble.R
\name{summary.survensemble_cv}
\alias{summary.survensemble_cv}
\title{Prints a summary of survensemble_cv object}
\usage{
\method{summary}{survensemble_cv}(object, ...)
}
\arguments{
\item{object}{survensemble_cv object}

\item{...}{additional arguments to be passed}
}
\value{
object
}
\description{
Prints a summary of survensemble_cv object
}
