% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survensemble.R
\name{summary.survensembleCV}
\alias{summary.survensembleCV}
\title{Prints a summary of survensembleCV object}
\usage{
\method{summary}{survensembleCV}(object, ...)
}
\arguments{
\item{object}{survensembleCV object}

\item{...}{additional arguments to be passed}
}
\value{
object
}
\description{
Prints a summary of survensembleCV object
}
