% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_cox_coxlasso.R
\name{survcox_cv}
\alias{survcox_cv}
\title{Cross-validates Cox or CoxLasso model}
\usage{
survcox_cv(
  df,
  predict.factors,
  fixed_time = NaN,
  cv_number = 5,
  randomseed = NULL,
  useCoxLasso = FALSE,
  parallel = FALSE,
  return_models = FALSE,
  repeat_cv = NULL
)
}
\arguments{
\item{df}{data frame with the data, "time" and "event" for survival outcome}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{at which event probabilities are computed}

\item{cv_number}{k in k-fold CV}

\item{randomseed}{random seed}

\item{useCoxLasso}{TRUE/FALSE}

\item{parallel}{TRUE/FALSE}

\item{return_models}{TRUE/FALSE, whether to return all CV models. Default is FALSE}

\item{repeat_cv}{if not NULL, repeats CV repeat_cv times}
}
\value{
output list: output$train, test, testaverage, traintaverage, time,tuned_cv_models
}
\description{
Cross-validates Cox or CoxLasso model
}
\examples{
df<- simsurv_nonlinear()
}
