% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_cox_coxlasso.R
\name{survcoxlasso_train}
\alias{survcoxlasso_train}
\title{Trains CoxLasso, using cv.glmnet(s="lambda.min")}
\usage{
survcoxlasso_train(
  df_train,
  predict.factors,
  fixed_time = NaN,
  retrain_cox = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df_train}{data frame with the data, "time" and "event" should describe survival outcome}

\item{predict.factors}{list of the column names to be used as predictors}

\item{fixed_time}{not used here, to re-align with other methods}

\item{retrain_cox}{whether to re-train coxph on non-zero predictors; FALSE by default}

\item{verbose}{TRUE/FALSE prints warnings if no predictors in Lasso}
}
\value{
fitted CoxPH model
}
\description{
Trains CoxLasso, using cv.glmnet(s="lambda.min")
}
