% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDeepHit.R
\name{survdeephit_predict}
\alias{survdeephit_predict}
\title{Computes event probabilities from a trained DeepHit}
\usage{
survdeephit_predict(
  trained_model,
  newdata,
  fixed_time,
  predict.factors,
  extrapsurvival = TRUE
)
}
\arguments{
\item{newdata}{new data for which predictions are made}

\item{fixed_time}{time of interest for which event probabilities are computed}

\item{predict.factors}{list of predictor names}

\item{extrapsurvival}{if probabilities are extrapolated beyond trained times (constant)}

\item{trained_object}{a trained model, output of survdhens_train()}
}
\value{
vector of predicted event probabilities
}
\description{
Computes event probabilities from a trained DeepHit
}
