% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDeepHit.R
\name{survdeephit_tune}
\alias{survdeephit_tune}
\title{Internal function for deephit_tune(), performs 1 CV}
\usage{
survdeephit_tune(
  df_tune,
  predict.factors,
  repeat_tune = 1,
  fixed_time = NaN,
  tuningparams = list(),
  max_grid_size = 10,
  inner_cv = 3,
  randomseed = NaN
)
}
\arguments{
\item{df_tune}{data, "time" and "event" should describe survival outcome}

\item{predict.factors}{list of predictor names}

\item{repeat_tune}{number of repeated CV for tuning (1 is default)}

\item{fixed_time}{time at which performance is maximized}

\item{tuningparams}{if given, list of hyperparameters, list(mod_alpha=c(), ...), otherwise a wide default grid is used}

\item{max_grid_size}{number of random grid searches for model tuning}

\item{inner_cv}{number of cross-validation folds for hyperparameters' tuning}

\item{randomseed}{random seed to control tuning including data splits}
}
\value{
list of outputs
}
\description{
Internal function for deephit_tune(), performs 1 CV
}
