% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDeepHit_ens.R
\name{survdhens_predict}
\alias{survdhens_predict}
\title{Computes event probabilities from a trained ensemble of CoxPH and DeepHit}
\usage{
survdhens_predict(
  trained_object,
  newdata,
  fixed_time,
  predict.factors,
  extrapsurvival = TRUE
)
}
\arguments{
\item{trained_object}{a trained model, output of survdhens_train()}

\item{newdata}{new data for which predictions are made}

\item{fixed_time}{time of interest for which event probabilities are computed}

\item{predict.factors}{list of predictor names}

\item{extrapsurvival}{if probabilities are extrapolated beyond trained times (constant)}
}
\value{
vector of predicted event probabilities
}
\description{
Computes event probabilities from a trained ensemble of CoxPH and DeepHit
}
