% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDeepHit_stack.R
\name{survdhstack_cv}
\alias{survdhstack_cv}
\title{Cross-validates Stacked Ensemble of CoxPH and DeepHit}
\usage{
survdhstack_cv(
  df,
  predict.factors,
  fixed_time = NaN,
  outer_cv = 3,
  inner_cv = 3,
  repeat_cv = 2,
  randomseed = NaN,
  return_models = FALSE,
  useCoxLasso = FALSE,
  tuningparams = list(),
  max_grid_size = 10,
  parallel = FALSE,
  verbose = FALSE,
  package_path = NaN,
  python_path = NaN
)
}
\arguments{
\item{df}{data, "time" and "event" should describe survival outcome}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{time at which performance is maximized}

\item{outer_cv}{number of cross-validation folds for model validation}

\item{inner_cv}{number of cross-validation folds for hyperparameters' tuning}

\item{repeat_cv}{number of cross-validation repetitions}

\item{randomseed}{random seed to control tuning including data splits}

\item{return_models}{TRUE/FALSE, if TRUE returns all trained models}

\item{useCoxLasso}{if CoxLasso is used (TRUE) or not (FALSE, default)}

\item{tuningparams}{if given, list of hyperparameters, list(mtry=c(), nodedepth=c(),nodesize=c()), otherwise a wide default grid is used
e.g. (mod_alpha = c(0.2,0.5),dropout= c(0.1,0.8), learning_rate= c(0.001,0.01), num_nodes = list(c(64,64)), epochs = 10, sigma = c(0.1,1,10), cuts =50,batch_size=50, early_stopping = FALSE, weight_decay= 0)}

\item{max_grid_size}{number of random grid searches for model tuning}

\item{parallel}{if parallel calculations are used}

\item{verbose}{FALSE(default)/TRUE}

\item{package_path}{survcompare package path if not installed as a library}

\item{python_path}{python path for survivalmodels}
}
\description{
Cross-validates Stacked Ensemble of CoxPH and DeepHit
}
