% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDeepHit_stack.R
\name{survdhstack_predict}
\alias{survdhstack_predict}
\title{Computes event probabilities from a trained stacked ensemble of DeepHit and CoxPH}
\usage{
survdhstack_predict(
  trained_object,
  newdata,
  fixed_time,
  predict.factors,
  extrapsurvival = FALSE
)
}
\arguments{
\item{trained_object}{a trained model, output of survdhstack_train()}

\item{newdata}{new data for which predictions are made}

\item{fixed_time}{time of interest for which event probabilities are computed}

\item{predict.factors}{list of predictor names}

\item{extrapsurvival}{if probabilities are extrapolated beyond trained times (constant)}
}
\value{
vector of predicted event probabilities
}
\description{
Computes event probabilities from a trained stacked ensemble of DeepHit and CoxPH
}
