% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_1.R
\name{survensemble1_predict}
\alias{survensemble1_predict}
\title{Predicts event probability for a fitted Ensemble 1 method}
\usage{
survensemble1_predict(
  model_1a,
  df_test,
  fixed_time,
  oob = FALSE,
  useCoxLasso = FALSE
)
}
\arguments{
\item{model_1a}{trained model}

\item{df_test}{test data}

\item{fixed_time}{time for which probabilities are computed}

\item{oob}{TRUE/FALSE , default is FALSE}

\item{useCoxLasso}{TRUE/FALSE , default is FALSE}
}
\value{
returns matrix(nrow = length(newdata), ncol = length(times))
}
\description{
\link[ensemblesurv:survcox_predict]{ensemblesurv:survcox_predict}
}
\examples{
df <- simsurv_nonlinear()
}
