% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ensemble.R
\name{survensemble_cv}
\alias{survensemble_cv}
\title{Cross-validates predictive performance for Ensemble 1}
\usage{
survensemble_cv(
  df,
  predict.factors,
  fixed_time = NaN,
  cv_number = 3,
  inner_cv = 3,
  repeat_cv = 2,
  randomseed = NULL,
  return_models = FALSE,
  useCoxLasso = FALSE,
  srf_tuning = NULL,
  oob = TRUE
)
}
\arguments{
\item{df}{data frame with the data, "time" and "event" for survival outcome}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{at which performance metrics are computed}

\item{cv_number}{k in k-fold CV, default 3}

\item{inner_cv}{kk in the inner look of kk-fold CV, default 3}

\item{repeat_cv}{if NULL, runs once (or 1), otherwise repeats CV}

\item{randomseed}{random seed}

\item{return_models}{TRUE/FALSE, if TRUE returns all CV objects}

\item{useCoxLasso}{TRUE/FALSE, default is FALSE}

\item{srf_tuning}{list of mtry, nodedepth, nodesize to tune, default is NULL}

\item{oob}{TRUE/FALSE use out-of-bag predictions while tuning instead of cross-validation, TRUE by default}
}
\value{
output list: output$train, test, testaverage, traintaverage, time
}
\description{
Cross-validates predictive performance for Ensemble 1
}
\examples{
df <- simsurv_nonlinear()
survensemble_cv(df, names(df)[1:4])
}
