% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survensemble.R
\name{survensembleCV}
\alias{survensembleCV}
\title{Cross-validates predictive performance for Ensemble 1}
\usage{
survensembleCV(
  df,
  predict.factors,
  fixed_time = NaN,
  cv_number = 3,
  inner_cv = 3,
  randomseed = NULL,
  useCoxLasso = FALSE,
  return_models = FALSE,
  srf_tuning = NULL,
  repeat_cv = NULL
)
}
\arguments{
\item{df}{data frame with the data, "time" and "event" for survival outcome}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{at which performance metrics are computed}

\item{cv_number}{k in k-fold CV, default 3}

\item{inner_cv}{kk in the inner look of kk-fold CV, default 3}

\item{randomseed}{random seed}

\item{useCoxLasso}{TRUE/FALSE, default is FALSE}

\item{return_models}{TRUE/FALSE, if TRUE returns all CV objects}

\item{srf_tuning}{list of mtry, nodedepth, nodesize to tune, default is NULL}

\item{repeat_cv}{if NULL, runs once, otherwise repeats CV}
}
\value{
output list: output$train, test, testaverage, traintaverage, time
}
\description{
Cross-validates predictive performance for Ensemble 1
}
\examples{
df <- simsurv_nonlinear()
}
