% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{survival_prob_km}
\alias{survival_prob_km}
\title{Calculates survival probability estimated by Kaplan-Meier survival curve
Uses polynomial extrapolation in survival function space, using poly(n=3)}
\usage{
survival_prob_km(df_km_train, times, estimate_censoring = FALSE)
}
\arguments{
\item{df_km_train}{event probabilities (!not survival)}

\item{times}{times at which survival is estimated}

\item{estimate_censoring}{FALSE by default, if TRUE, event and censoring is reversed (for IPCW calculations)}
}
\value{
vector of survival probabilities for time_points
}
\description{
Calculates survival probability estimated by Kaplan-Meier survival curve
Uses polynomial extrapolation in survival function space, using poly(n=3)
}
