% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_survivalrandomforest.R
\name{survrf_cv}
\alias{survrf_cv}
\title{Cross-validates SRF model}
\usage{
survrf_cv(
  df,
  predict.factors,
  fixed_time = NaN,
  cv_number = 3,
  inner_cv = 3,
  randomseed = NULL,
  srf_tuning = NULL,
  parallel = FALSE,
  return_models = FALSE,
  repeat_cv = NULL
)
}
\arguments{
\item{df}{data frame with the data, "time" and "event" for survival outcome}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{at which performance metrics are computed}

\item{cv_number}{k in k-fold CV, default 3}

\item{inner_cv}{kk in the inner look of kk-fold CV, default 3}

\item{randomseed}{random seed}

\item{srf_tuning}{list of mtry, nodedepth and nodesize, default is NULL}

\item{parallel}{TRUE/FALSE, default = FALSE}

\item{return_models}{TRUE/FALSE, if TRUE returns all CV objects}

\item{repeat_cv}{if NULL, runs once, otherwise repeats CV}
}
\value{
output list: output$train, test, testaverage, traintaverage, time
}
\description{
Cross-validates SRF model
}
\examples{
df <- simsurv_nonlinear()
}
