% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_survivalrandomforest.R
\name{survrf_predict}
\alias{survrf_predict}
\title{Predicts event probability for a fitted SRF model}
\usage{
survrf_predict(model_srf, df_test, fixed_time, oob = FALSE)
}
\arguments{
\item{model_srf}{trained model}

\item{df_test}{test data}

\item{fixed_time}{time for which probabilities are computed}

\item{oob}{TRUE/FALSE , default is FALSE}
}
\value{
returns matrix(nrow = length(newdata), ncol = length(times))
}
\description{
Predicts event probability for a fitted SRF model randomForestSRC::rfsrc.
Essentially a wrapper of \link{srf_survival_prob_for_time}.
}
\examples{
df <- simsurv_nonlinear()
}
