% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_SRF.R
\name{survsrf_cv}
\alias{survsrf_cv}
\title{Cross-validates SRF model}
\usage{
survsrf_cv(
  df,
  predict.factors,
  fixed_time = NaN,
  cv_number = 3,
  repeat_cv = 2,
  randomseed = NULL,
  return_models = FALSE,
  inner_cv = 3,
  srf_tuning = NULL,
  oob = TRUE
)
}
\arguments{
\item{df}{data frame with the data, "time" and "event" for survival outcome}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{at which performance metrics are computed}

\item{cv_number}{k in k-fold CV, default 3}

\item{repeat_cv}{if NULL, runs once, otherwise repeats CV}

\item{randomseed}{random seed}

\item{return_models}{TRUE/FALSE, if TRUE returns all CV objects}

\item{inner_cv}{k in the inner loop of k-fold CV, default 3}

\item{srf_tuning}{list of candidate mtry, nodedepth and nodesize, default is NULL}

\item{oob}{TRUE/FALSE use out-of-bag prediction accuracy while tuning instead of cross-validation, TRUE by defaul}
}
\value{
output list: output$train, test, testaverage, traintaverage, time
}
\description{
Cross-validates SRF model
}
\examples{
df <- simulate_nonlinear(100)
survsrf_cv(df, names(df)[1:4])
}
