% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSRF.R
\name{survsrf_predict}
\alias{survsrf_predict}
\title{Predicts event probability by a trained Survival Random Forest}
\usage{
survsrf_predict(trained_model, newdata, fixed_time, extrapsurvival = TRUE)
}
\arguments{
\item{trained_model}{a trained SRF model, output of survsrf_train(), or randomForestSRC::rfsrc()}

\item{newdata}{new data for which predictions are made}

\item{fixed_time}{time of interest for which event probabilities are computed}

\item{extrapsurvival}{if probabilities are extrapolated beyond trained times (using probability of
the lastest available time). Can be helpful for cross-validation of small data, where random
split may cause the time of interest being outside of the training set.}
}
\value{
vector of predicted event probabilities
}
\description{
Predicts event probability by a trained Survival Random Forest
}
