% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSRF.R
\name{survsrf_tune_single}
\alias{survsrf_tune_single}
\title{Internal function for survsrf_tune(), performs 1 CV}
\usage{
survsrf_tune_single(
  df_tune,
  predict.factors,
  fixed_time = NaN,
  grid_hyperparams = c(),
  inner_cv = 3,
  randomseed = NaN,
  progressbar = FALSE
)
}
\arguments{
\item{df_tune}{data}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{predictions for which time are computed for c-index}

\item{grid_hyperparams}{hyperparameters grid (or a default will be used )}

\item{inner_cv}{number of folds for each CV}

\item{randomseed}{randomseed}

\item{progressbar}{FALSE(default)/TRUE}
}
\value{
output=list(grid, cindex, cindex_mean)
}
\description{
Internal function for survsrf_tune(), performs 1 CV
}
