% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSRF_ens.R
\name{survsrfens_predict}
\alias{survsrfens_predict}
\title{Predicts event probability by a trained sequential ensemble of Survival Random Forest and CoxPH}
\usage{
survsrfens_predict(trained_model, newdata, fixed_time, extrapsurvival = TRUE)
}
\arguments{
\item{trained_model}{a trained model, output of survsrfens_train()}

\item{newdata}{new data for which predictions are made}

\item{fixed_time}{time of interest, for which event probabilities are computed}

\item{extrapsurvival}{if probabilities are extrapolated beyond trained times (constant)}
}
\value{
vector of predicted event probabilities
}
\description{
Predicts event probability by a trained sequential ensemble of Survival Random Forest and CoxPH
}
