% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSRF_stack.R
\name{survsrfstack_predict}
\alias{survsrfstack_predict}
\title{Predicts event probability by a trained stacked ensemble of Survival Random Forest and CoxPH}
\usage{
survsrfstack_predict(
  trained_object,
  newdata,
  fixed_time,
  predict.factors,
  extrapsurvival = TRUE
)
}
\arguments{
\item{trained_object}{a trained model, output of survsrfstack_train()}

\item{newdata}{new data for which predictions are made}

\item{fixed_time}{time of interest, for which event probabilities are computed}

\item{predict.factors}{list of predictor names}

\item{extrapsurvival}{if probabilities are extrapolated beyond trained times (constant)}
}
\value{
vector of predicted event probabilities
}
\description{
Predicts event probability by a trained stacked ensemble of Survival Random Forest and CoxPH
}
