% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{survval}
\alias{survval}
\title{Computes performance statistics given the predicted event probabilities}
\usage{
survval(
  y_predict,
  times_to_predict,
  df_train,
  df_test,
  weighted = TRUE,
  alpha = "logit"
)
}
\arguments{
\item{y_predict}{probabilities of event by times_to_predict (matrix=observations x times)}

\item{times_to_predict}{times for which event probabilities are given}

\item{df_train}{train data, data frame}

\item{df_test}{test data, data frame}

\item{weighted}{TRUE/FALSE, for IPWC}

\item{alpha}{calibration alpha as mean difference or from logistic regression}
}
\value{
data.frame(T, AUCROC, Brier Score, Scaled Brier Score, C_score, Calib slope, Calib alpha)
}
\description{
Computes performance statistics given the predicted event probabilities
}
