% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSRF.R
\name{ml_hyperparams_srf}
\alias{ml_hyperparams_srf}
\title{Internal function for getting grid of hyperparameters
for random or grid search of size = max_grid_size}
\usage{
ml_hyperparams_srf(
  mlparams = list(),
  p = 10,
  max_grid_size = 10,
  dftune_size = 1000,
  randomseed = NaN
)
}
\arguments{
\item{mlparams}{list of params}

\item{p}{number of predictors to detine mtry options}

\item{max_grid_size}{grid size for tuning}

\item{dftune_size}{size of the tuning data to define nodesize options}

\item{randomseed}{randomseed to select the tuning grid}
}
\description{
Internal function for getting grid of hyperparameters
for random or grid search of size = max_grid_size
}
