% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ensemble.R
\name{predict.survensemble}
\alias{predict.survensemble}
\title{Predicts event probability for a fitted survensemble}
\usage{
\method{predict}{survensemble}(object, newdata, fixed_time, oob = FALSE, ...)
}
\arguments{
\item{object}{trained survensemble model}

\item{newdata}{test data}

\item{fixed_time}{time for which probabilities are computed}

\item{oob}{TRUE/FALSE , default is FALSE, if out of bag predictions are to be made from SRF}

\item{...}{other parameters to pass}
}
\value{
matrix of predictions for observations in newdata by times
}
\description{
\link[survcompare:predict.survensemble]{predict.survensemble}
}
