% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survcompare.R
\name{print.survcompare}
\alias{print.survcompare}
\title{Prints survcompare object}
\usage{
\method{print}{survcompare}(obj, ...)
}
\arguments{
\item{obj}{a surcompare object, output of a survcompare() function}
}
\description{
Prints survcompare object
}
