% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survcompare.R
\name{summary.survcompare}
\alias{summary.survcompare}
\title{Prints summary for a survcompare object}
\usage{
\method{summary}{survcompare}(output_object, ...)
}
\arguments{
\item{output_object}{a survcompare object, output of a survcompare() function}
}
\description{
Prints summary for a survcompare object
}
