% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcompare.R
\name{survcompare}
\alias{survcompare}
\title{Cross-validates predictive performance of the baseline Cox-PH model and Survival random forest}
\usage{
survcompare(
  df_train,
  predict_factors,
  predict_t = NULL,
  randomseed = NULL,
  useCoxLasso = FALSE,
  outer_cv = 5,
  inner_cv = 3,
  srf_tuning = NaN,
  return_models = FALSE,
  repeat_cv = NULL,
  train_srf = FALSE
)
}
\arguments{
\item{df_train}{data frame}

\item{predict_factors}{list of predictors to use}

\item{predict_t}{prediction time of interest. If NULL, 90\% quantile of event times is used}

\item{randomseed}{random seed for replication}

\item{useCoxLasso}{TRUE / FALSE, FALSE by default}

\item{outer_cv}{k in k-fold CV}

\item{inner_cv}{k in k-fold CV for internal tuning of survival random forest}

\item{srf_tuning}{list of tuning parameters for srf, e.g. list("mtry"=c(3,6,16), "ndepth" = c(3,6,19))}

\item{return_models}{TRUE/FALSE, if returns all the trained models; default is FALSE, only performance is returned}

\item{repeat_cv}{if NULL, runs once, otherwise repeats several times, reports average of all}

\item{train_srf}{TRUE/FALSE, default is FALSE, as Ensemble 1 by definition is very similar}
}
\value{
outcome = list(data frame with performance results, fitted Cox models, fitted SRF)
}
\description{
The following models are evaluated:
\enumerate{
\item Cox-PH (survival package, survival::coxph)
\item Cox-Lasso (glmnet package, glmnet::cox.fit)
\item Survival random forest, SRF (randomForestSRC::rfsrc)
\item Ensemble of Cox-PH or Cox-Lasso and SRF
}

The same random seed for the train/test splits are used for all models.
The comparison can aid model selection and quantify the loss of predictive
performance (if any) if Cox-PH is used instead of a more complex model such as SRF
The difference in performance of the Ensembled Cox and SRF and the baseline Cox-PH
can be viewed as quantification of the non-linear and cross-terms contribution to
the predictive power of the supplied predictors
}
