% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_CoxPH.R
\name{survcox_cv}
\alias{survcox_cv}
\title{Cross-validates Cox or CoxLasso model}
\usage{
survcox_cv(
  df,
  predict.factors,
  fixed_time = NaN,
  cv_number = 3,
  repeat_cv = 2,
  randomseed = NULL,
  return_models = FALSE,
  inner_cv = 3,
  useCoxLasso = FALSE
)
}
\arguments{
\item{df}{data frame with the data, "time" and "event" for survival outcome}

\item{predict.factors}{list of predictor names}

\item{fixed_time}{at which performance metrics are computed}

\item{cv_number}{k in k-fold CV, default 3}

\item{repeat_cv}{if NULL, runs once, otherwise repeats CV}

\item{randomseed}{random seed}

\item{return_models}{TRUE/FALSE, if TRUE returns all CV objects}

\item{inner_cv}{k in the inner loop of k-fold CV, default 3}

\item{useCoxLasso}{TRUE/FALSE, FALSE by default}
}
\value{
output list: output$train, test, testaverage, traintaverage, time,tuned_cv_models
}
\description{
Cross-validates Cox or CoxLasso model
}
\examples{
df<- simulate_nonlinear()
}
