% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_cox_coxlasso.R
\name{survcox_predict}
\alias{survcox_predict}
\title{Computes event probabilities from a trained cox model}
\usage{
survcox_predict(model_cox, newdata, times)
}
\arguments{
\item{model_cox}{pre-trained cox model}

\item{newdata}{data to compute event probabilities for}

\item{times}{at which event probabilities are computed}
}
\value{
returns matrix(nrow = length(newdata), ncol = length(times))
}
\description{
Computes event probabilities from a trained cox model
}
\examples{
df<- simsurv_nonlinear()
}
