% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_cox_coxlasso.R
\name{survcox_train}
\alias{survcox_train}
\title{Trains CoxPH using survival package, or trains CoxLasso (cv.glmnet, lambda.min),
and then re-trains survival:coxph on non-zero predictors}
\usage{
survcox_train(
  df_train,
  predict.factors,
  useCoxLasso = FALSE,
  fixed_time = NaN,
  retrain_cox = FALSE
)
}
\arguments{
\item{df_train}{data, "time" and "event" should describe survival outcome}

\item{predict.factors}{list of the column names to be used as predictors}

\item{useCoxLasso}{TRUE or FALSE}

\item{fixed_time}{not used here, to re-align with other methods}

\item{retrain_cox}{if useCoxLasso is TRUE, whether to re-train coxph on non-zero predictors, FALSE by default}
}
\value{
fitted CoxPH or CoxLasso model
}
\description{
Trains CoxPH using survival package, or trains CoxLasso (cv.glmnet, lambda.min),
and then re-trains survival:coxph on non-zero predictors
}
\examples{
df<- simsurv_nonlinear()
}
