% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_SRF.R
\name{survsrf_predict}
\alias{survsrf_predict}
\title{Predicts event probability for a fitted SRF model}
\usage{
survsrf_predict(trained_model, newdata, fixed_time, oob = FALSE)
}
\arguments{
\item{trained_model}{trained model}

\item{newdata}{test data}

\item{fixed_time}{time for which probabilities are computed}

\item{oob}{TRUE/FALSE for out-of-bag predictions, default is FALSE}
}
\value{
returns vector of predictions (or matrix if fixed_time is a vector of times)
}
\description{
Predicts event probability for a fitted SRF model randomForestSRC::rfsrc.
Essentially a wrapper of \link{srf_survival_prob_for_time}.
}
